requirements.txt
blis==0.7.4
cachetools==4.2.0
cassandra-driver==3.24.0
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
fastai==2.2.5
fastcore==1.3.19
fastprogress==1.0.0
Flask==1.1.2
geomet==0.2.1.post1
google-api-core==1.25.0
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
murmurhash==1.0.5
numpy==1.19.5
oauthlib==3.1.0
packaging==20.8
pandas==1.2.1
Pillow==8.1.0
plac==1.1.3
preshed==3.0.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn==0.24.1
scipy==1.6.0
simplejson==3.17.2
six==1.15.0
spacy==2.3.5
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
